test: split oversized transformer suites
