+++ tests/unit/responses-transformer-streaming.test.ts
@@ -22,7 +22,7 @@ import test from "node:test";
 import assert from "node:assert/strict";
 
 const { createResponsesApiTransformStream } =
-  await import("../../../open-sse/transformer/responsesTransformer.ts");
+  await import("../../open-sse/transformer/responsesTransformer.ts");
 
 const encoder = new TextEncoder();
 const decoder = new TextDecoder();
@@ -132,399 +132,15 @@ function getCreated(output: string): Record<string, unknown> {
 
 // ───────────────────────── 1. tool_use blocks ─────────────────────────
 
-test("tool_use is NOT emitted when no Anthropic-shaped tools and no opt-in", async () => {
-  const output = await runTransformStream([
-    sse({ 
-      toolCalls: [{ index: 0, id: "call_a", function: { name: "lookup", arguments: "{}" } }],
-    }),
-    sse({ finishReason: "tool_calls" }),
-  ]);
-  const events = parseSseOutput(output);
-  const outputItemAdded = events.filter((e) => e.event === "response.output_item.added");
-  assert.equal(outputItemAdded.length, 1);
-  const item = JSON.parse(outputItemAdded[0].data!).item as Record<string, unknown>;
-  assert.equal(item.type, "function_call");
-});
-
-test("tool_use IS emitted automatically when tools[] contains Anthropic-shaped entries", async () => {
-  const tools = [
-    {
-      name: "lookup",
-      description: "Look up a record",
-      input_schema: {
-        type: "object",
-        properties: { q: { type: "string" } },
-        required: ["q"],
-      },
-    },
-  ];
-  const output = await runTransformStream(
-    [
-      sse({ 
-        toolCalls: [
-          {
-            index: 0,
-            id: "call_b",
-            function: { name: "lookup", arguments: JSON.stringify({ q: "hi" }) },
-          },
-        ],
-      }),
-      sse({ finishReason: "tool_calls" }),
-    ],
-    { tools }
-  );
-  const events = parseSseOutput(output);
-
-  const added = events.filter((e) => e.event === "response.output_item.added");
-  assert.equal(added.length, 2, "one function_call + one parallel tool_use output_item.added");
-  const types = added.map((e) => (JSON.parse(e.data!).item as { type: string }).type);
-  assert.deepEqual(types.sort(), ["function_call", "tool_use"]);
-
-  const toolUseAdded = added.find(
-    (e) => (JSON.parse(e.data!).item as { type: string }).type === "tool_use"
-  );
-  const toolUseItem = JSON.parse(toolUseAdded!.data!).item as Record<string, unknown>;
-  assert.equal(toolUseItem.name, "lookup");
-  assert.deepEqual(toolUseItem.input, {});
-  assert.equal(toolUseItem.id, "fc_call_b", "tool_use shares id with the function_call");
-});
-
-test("tool_use done event has parsed input when arguments are valid JSON", async () => {
-  const tools = [
-    {
-      name: "lookup",
-      input_schema: { type: "object", properties: { q: { type: "string" } } },
-    },
-  ];
-  const args = JSON.stringify({ q: "hello", n: 2 });
-  const output = await runTransformStream(
-    [
-      sse({ toolCalls: [{ index: 0, id: "call_c", function: { name: "lookup", arguments: args } }] }),
-      sse({ finishReason: "tool_calls" }),
-    ],
-    { tools }
-  );
-
-  const completed = getCompleted(output);
-  const outputArr = completed.output as Array<Record<string, unknown>>;
-  const funcCall = outputArr.find((i) => i.type === "function_call");
-  const toolUse = outputArr.find((i) => i.type === "tool_use");
-  assert.ok(funcCall, "function_call item present");
-  assert.ok(toolUse, "tool_use item present");
-  assert.equal(toolUse.id, funcCall.id, "tool_use and function_call share id");
-  assert.deepEqual(
-    toolUse.input,
-    { q: "hello", n: 2 },
-    "input is the parsed JSON object, not a string"
-  );
-  assert.equal(typeof funcCall.arguments, "string");
-  assert.equal(typeof toolUse.input, "object");
-});
-
-test("tool_use input is empty-object when arguments are invalid JSON", async () => {
-  const tools = [{ name: "broken", input_schema: { type: "object" } }];
-  const output = await runTransformStream(
-    [
-      sse({ 
-        toolCalls: [
-          { index: 0, id: "call_d", function: { name: "broken", arguments: "not json" } },
-        ],
-      }),
-      sse({ finishReason: "tool_calls" }),
-    ],
-    { tools }
-  );
-  const completed = getCompleted(output);
-  const toolUse = (completed.output as Array<Record<string, unknown>>).find(
-    (i) => i.type === "tool_use"
-  );
-  assert.ok(toolUse);
-  assert.deepEqual(toolUse.input, {}, "invalid JSON falls back to empty object");
-});
-
-test("emitToolUse=true forces tool_use emission even for Chat-shaped tools", async () => {
-  const tools = [{ type: "function", function: { name: "lookup", parameters: {} } }];
-  const output = await runTransformStream(
-    [
-      sse({ 
-        toolCalls: [{ index: 0, id: "call_e", function: { name: "lookup", arguments: "{}" } }],
-      }),
-      sse({ finishReason: "tool_calls" }),
-    ],
-    { tools, emitToolUse: true }
-  );
-  const completed = getCompleted(output);
-  const outputArr = completed.output as Array<Record<string, unknown>>;
-  assert.ok(outputArr.find((i) => i.type === "function_call"), "function_call still present");
-  assert.ok(outputArr.find((i) => i.type === "tool_use"), "tool_use forced on by opt-in");
-});
-
-test("tool_use and function_call share the SAME id and output_index in dense output", async () => {
-  const tools = [{ name: "lookup", input_schema: { type: "object" } }];
-  const output = await runTransformStream(
-    [
-      sse({ 
-        toolCalls: [{ index: 0, id: "call_f", function: { name: "lookup", arguments: "{}" } }],
-      }),
-      sse({ finishReason: "tool_calls" }),
-    ],
-    { tools }
-  );
-  const completed = getCompleted(output);
-  const outputArr = completed.output as Array<Record<string, unknown>>;
-  const fc = outputArr.find((i) => i.type === "function_call") as Record<string, unknown>;
-  const tu = outputArr.find((i) => i.type === "tool_use") as Record<string, unknown>;
-  assert.equal(fc.id, tu.id);
-});
-
-test("tool_use streaming events: output_item.added + output_item.done both fire", async () => {
-  const tools = [{ name: "lookup", input_schema: { type: "object" } }];
-  const output = await runTransformStream(
-    [
-      sse({ 
-        toolCalls: [{ index: 0, id: "call_g", function: { name: "lookup", arguments: "{}" } }],
-      }),
-      sse({ finishReason: "tool_calls" }),
-    ],
-    { tools }
-  );
-  const events = parseSseOutput(output);
-  const addedTypes = events
-    .filter((e) => e.event === "response.output_item.added")
-    .map((e) => (JSON.parse(e.data!).item as { type: string }).type);
-  const doneTypes = events
-    .filter((e) => e.event === "response.output_item.done")
-    .map((e) => (JSON.parse(e.data!).item as { type: string }).type);
-  assert.deepEqual(addedTypes.sort(), ["function_call", "tool_use"], "both added");
-  assert.deepEqual(doneTypes.sort(), ["function_call", "tool_use"], "both done");
-});
-
-// ───────────────────────── 2. structured_outputs ─────────────────────────
-
-test("structured_outputs: response_format json_schema is echoed on response", async () => {
-  const request = {
-    model: "gpt-5.5",
-    input: "extract",
-    response_format: {
-      type: "json_schema",
-      json_schema: {
-        name: "user_profile",
-        schema: {
-          type: "object",
-          properties: { name: { type: "string" } },
-          required: ["name"],
-        },
-        strict: true,
-      },
-    },
-  };
-  const output = await runTransformStream(
-    [sse({ content: JSON.stringify({ name: "K" }) }), sse({ finishReason: "stop" })],
-    { request }
-  );
-
-  const created = getCreated(output);
-  const completed = getCompleted(output);
-  const expectedFormat = {
-    type: "json_schema",
-    name: "user_profile",
-    schema: {
-      type: "object",
-      properties: { name: { type: "string" } },
-      required: ["name"],
-    },
-    strict: true,
-  };
-  assert.deepEqual(created.output_format, expectedFormat);
-  assert.deepEqual(completed.output_format, expectedFormat);
-});
-
-test("structured_outputs: json_object is echoed on response", async () => {
-  const request = { model: "gpt-5.5", input: "x", response_format: { type: "json_object" } };
-  const output = await runTransformStream(
-    [sse({ content: "{}" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const completed = getCompleted(output);
-  assert.deepEqual(completed.output_format, { type: "json_object" });
-});
-
-test("structured_outputs: text format is echoed on response", async () => {
-  const request = { model: "gpt-5.5", input: "x", response_format: { type: "text" } };
-  const output = await runTransformStream(
-    [sse({ content: "hi" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const completed = getCompleted(output);
-  assert.deepEqual(completed.output_format, { type: "text" });
-});
-
-test("structured_outputs: absent response_format does NOT add output_format field", async () => {
-  const output = await runTransformStream([
-    sse({ content: "hi" }),
-    sse({ finishReason: "stop" }),
-  ]);
-  const completed = getCompleted(output);
-  assert.equal(
-    "output_format" in completed,
-    false,
-    "output_format should be absent when request has no response_format"
-  );
-});
-
-test("structured_outputs: json_schema without strict defaults strict=true", async () => {
-  const request = {
-    model: "gpt-5.5",
-    response_format: {
-      type: "json_schema",
-      json_schema: { name: "x", schema: { type: "object" } },
-    },
-  };
-  const output = await runTransformStream(
-    [sse({ content: "{}" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const completed = getCompleted(output);
-  const fmt = completed.output_format as Record<string, unknown>;
-  assert.equal(fmt.strict, true, "strict defaults to true when omitted");
-});
-
-test("structured_outputs: strict:false is honored", async () => {
-  const request = {
-    model: "gpt-5.5",
-    response_format: {
-      type: "json_schema",
-      json_schema: { name: "x", schema: {}, strict: false },
-    },
-  };
-  const output = await runTransformStream(
-    [sse({ content: "{}" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const completed = getCompleted(output);
-  const fmt = completed.output_format as Record<string, unknown>;
-  assert.equal(fmt.strict, false);
-});
-
-// ───────────────────────── 3. prompt_cache_control ─────────────────────────
-
-test("prompt_cache_key: snake_case from request is echoed on response", async () => {
-  const request = { model: "gpt-5.5", input: "x", prompt_cache_key: "user-42-session-7" };
-  const output = await runTransformStream(
-    [sse({ content: "hi" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const created = getCreated(output);
-  const completed = getCompleted(output);
-  assert.equal(created.prompt_cache_key, "user-42-session-7");
-  assert.equal(completed.prompt_cache_key, "user-42-session-7");
-});
-
-test("prompt_cache_key: camelCase alias is also accepted", async () => {
-  const request = { model: "gpt-5.5", input: "x", promptCacheKey: "session-99" };
-  const output = await runTransformStream(
-    [sse({ content: "hi" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const completed = getCompleted(output);
-  assert.equal(completed.prompt_cache_key, "session-99");
-});
-
-test("prompt_cache_key: snake_case wins when both forms are present", async () => {
-  const request = {
-    model: "gpt-5.5",
-    input: "x",
-    prompt_cache_key: "snake",
-    promptCacheKey: "camel",
-  };
-  const output = await runTransformStream(
-    [sse({ content: "hi" }), sse({ finishReason: "stop" })],
-    { request }
-  );
-  const completed = getCompleted(output);
-  assert.equal(completed.prompt_cache_key, "snake");
-});
-
-test("prompt_cache_key: absent when no cache key in request", async () => {
-  const output = await runTransformStream([
-    sse({ content: "hi" }),
-    sse({ finishReason: "stop" }),
-  ]);
-  const completed = getCompleted(output);
-  assert.equal("prompt_cache_key" in completed, false);
-});
-
-test("usage: cache_creation_input_tokens + cache_read_input_tokens are forwarded as-is", async () => {
-  const output = await runTransformStream([
-    sse({ content: "hi" }),
-    sse({
-      usageOnly: true,
-      usage: {
-        prompt_tokens: 10,
-        completion_tokens: 2,
-        total_tokens: 12,
-        cache_creation_input_tokens: 900,
-        cache_read_input_tokens: 1234,
-      },
-    }),
-  ]);
-  const completed = getCompleted(output);
-  const usage = completed.usage as Record<string, unknown>;
-  assert.equal(usage.cache_creation_input_tokens, 900);
-  assert.equal(usage.cache_read_input_tokens, 1234);
-  assert.equal(usage.prompt_tokens, 10);
-  assert.equal(usage.completion_tokens, 2);
-});
-
-test("usage: missing cache_* fields are NOT invented", async () => {
-  const output = await runTransformStream([
-    sse({ content: "hi" }),
-    sse({
-      usageOnly: true,
-      usage: { prompt_tokens: 10, completion_tokens: 2, total_tokens: 12 },
-    }),
-  ]);
-  const completed = getCompleted(output);
-  const usage = completed.usage as Record<string, unknown>;
-  assert.equal("cache_creation_input_tokens" in usage, false);
-  assert.equal("cache_read_input_tokens" in usage, false);
-});
-
-test("usage: OpenAI-shaped cached_tokens is also forwarded", async () => {
-  const output = await runTransformStream([
-    sse({ content: "hi" }),
-    sse({
-      usageOnly: true,
-      usage: {
-        prompt_tokens: 10,
-        completion_tokens: 2,
-        total_tokens: 12,
-        cached_tokens: 800,
-      },
-    }),
-  ]);
-  const completed = getCompleted(output);
-  const usage = completed.usage as Record<string, unknown>;
-  assert.equal(usage.cached_tokens, 800);
-});
-
-// ───────────────────────── 4. Existing behavior — baseline guards ─────────────────────────
-
 test("baseline: response.created is emitted first", async () => {
-  const output = await runTransformStream([
-    sse({ content: "x" }),
-    sse({ finishReason: "stop" }),
-  ]);
+  const output = await runTransformStream([sse({ content: "x" }), sse({ finishReason: "stop" })]);
   const events = parseSseOutput(output);
   const first = events.find((e) => e.event !== null);
   assert.equal(first.event, "response.created");
 });
 
 test("baseline: response.completed is the last named event before [DONE]", async () => {
-  const output = await runTransformStream([
-    sse({ content: "x" }),
-    sse({ finishReason: "stop" }),
-  ]);
+  const output = await runTransformStream([sse({ content: "x" }), sse({ finishReason: "stop" })]);
   const events = parseSseOutput(output);
   const last = events[events.length - 1];
   const penultimate = events[events.length - 2];
@@ -567,8 +183,14 @@ test("baseline: reasoning_content is emitted as a reasoning item", async () => {
   ]);
   const completed = getCompleted(output);
   const items = completed.output as Array<Record<string, unknown>>;
-  assert.ok(items.find((i) => i.type === "reasoning"), "reasoning item present");
-  assert.ok(items.find((i) => i.type === "message"), "message item present");
+  assert.ok(
+    items.find((i) => i.type === "reasoning"),
+    "reasoning item present"
+  );
+  assert.ok(
+    items.find((i) => i.type === "message"),
+    "message item present"
+  );
   const reasoningIdx = items.findIndex((i) => i.type === "reasoning");
   const messageIdx = items.findIndex((i) => i.type === "message");
   assert.ok(reasoningIdx < messageIdx, "reasoning precedes message in dense output");
@@ -592,8 +214,7 @@ test("baseline: <think>...</think> tags become reasoning items", async () => {
   const completed = getCompleted(output);
   const items = completed.output as Array<Record<string, unknown>>;
   const reasoning = items.find((i) => i.type === "reasoning") as
-    | { summary: Array<{ text: string }> }
-    | undefined;
+    { summary: Array<{ text: string }> } | undefined;
   const message = items.find((i) => i.type === "message");
   assert.ok(reasoning, "reasoning item present");
   assert.equal(reasoning!.summary[0].text, "plan");
@@ -649,7 +270,7 @@ test("baseline: dense output preserves ascending output_index across mixed messa
   // message (output_index 2+1 since reasoning is closed first).
   const output = await runTransformStream([
     sse({ index: 2, content: "later" }),
-    sse({ 
+    sse({
       index: 0,
       toolCalls: [{ index: 0, id: "call_d1", function: { name: "f", arguments: "{}" } }],
     }),
@@ -667,7 +288,10 @@ test("baseline: dense output preserves ascending output_index across mixed messa
 
 test("baseline: empty choices with usage-only chunks do not break the stream", async () => {
   // OpenAI sends a usage-only chunk at the end (no `choices`).
-  const usageOnly = sse({ usageOnly: true, usage: { prompt_tokens: 1, completion_tokens: 1, total_tokens: 2 } });
+  const usageOnly = sse({
+    usageOnly: true,
+    usage: { prompt_tokens: 1, completion_tokens: 1, total_tokens: 2 },
+  });
   const output = await runTransformStream([
     sse({ content: "x" }),
     usageOnly,
@@ -680,12 +304,12 @@ test("baseline: empty choices with usage-only chunks do not break the stream", a
 });
 
 test("baseline: response.error is not synthesized on finish — only response.completed fires for a clean stop", async () => {
-  const output = await runTransformStream([
-    sse({ content: "x" }),
-    sse({ finishReason: "stop" }),
-  ]);
+  const output = await runTransformStream([sse({ content: "x" }), sse({ finishReason: "stop" })]);
   const events = parseSseOutput(output);
-  assert.equal(events.find((e) => e.event === "response.error"), undefined);
+  assert.equal(
+    events.find((e) => e.event === "response.error"),
+    undefined
+  );
   assert.ok(events.find((e) => e.event === "response.completed"));
 });
 
@@ -740,9 +364,7 @@ test("baseline: tools[] with input_schema is detected even when parameters is al
   // Chat-style `parameters`. If both are present the entry is ambiguous and
   // the transformer falls back to function_call-only (the safer default — the
   // client can opt in via the explicit `emitToolUse: true` flag).
-  const tools = [
-    { name: "f", input_schema: { type: "object" }, parameters: { type: "object" } },
-  ];
+  const tools = [{ name: "f", input_schema: { type: "object" }, parameters: { type: "object" } }];
   const output = await runTransformStream(
     [
       sse({ toolCalls: [{ index: 0, id: "call_t2", function: { name: "f", arguments: "{}" } }] }),
@@ -759,7 +381,7 @@ test("baseline: tools[] with input_schema is detected even when parameters is al
 
 test("baseline: multiple tool_calls in different indexes are emitted in order", async () => {
   const output = await runTransformStream([
-    sse({ 
+    sse({
       toolCalls: [
         { index: 2, id: "call_b", function: { name: "b", arguments: "{}" } },
         { index: 0, id: "call_a", function: { name: "a", arguments: "{}" } },
@@ -784,10 +406,7 @@ test("baseline: response.completed output is always an array (never null/undefin
 });
 
 test("baseline: sequence_number increments across all emitted events", async () => {
-  const output = await runTransformStream([
-    sse({ content: "x" }),
-    sse({ finishReason: "stop" }),
-  ]);
+  const output = await runTransformStream([sse({ content: "x" }), sse({ finishReason: "stop" })]);
   const events = parseSseOutput(output).filter((e) => e.event !== null);
   const seqs = events.map((e) => {
     const data = JSON.parse(e.data!);
@@ -839,10 +458,7 @@ test("baseline: request without options behaves identically to legacy two-arg ca
 });
 
 test("baseline: response.completed always ends with status:completed", async () => {
-  const output = await runTransformStream([
-    sse({ content: "x" }),
-    sse({ finishReason: "stop" }),
-  ]);
+  const output = await runTransformStream([sse({ content: "x" }), sse({ finishReason: "stop" })]);
   const completed = getCompleted(output);
   assert.equal(completed.status, "completed");
   assert.equal(completed.object, "response");
